(1248) Split up shantytownModel
