+++ packages/api/server/models/shantytownModel/_common/SQL.js
@@ -1,0 +1,126 @@
+module.exports = {
+    selection: {
+        'shantytowns.shantytown_id': 'id',
+        'shantytowns.name': 'name',
+        'shantytowns.status': 'status',
+        'shantytowns.declared_at': 'declaredAt',
+        'shantytowns.built_at': 'builtAt',
+        'shantytowns.closed_at': 'closedAt',
+        'shantytowns.latitude': 'latitude',
+        'shantytowns.longitude': 'longitude',
+        'shantytowns.address': 'address',
+        'shantytowns.address_details': 'addressDetails',
+        '(SELECT regexp_matches(shantytowns.address, \'^(.+) [0-9]+ [^,]+,? [0-9]+,? [^, ]+(,.+)?$\'))[1]': 'addressSimple',
+        'shantytowns.population_total': 'populationTotal',
+        'shantytowns.population_couples': 'populationCouples',
+        'shantytowns.population_minors': 'populationMinors',
+        'shantytowns.population_minors_0_3': 'populationMinors0To3',
+        'shantytowns.population_minors_3_6': 'populationMinors3To6',
+        'shantytowns.population_minors_6_12': 'populationMinors6To12',
+        'shantytowns.population_minors_12_16': 'populationMinors12To16',
+        'shantytowns.population_minors_16_18': 'populationMinors16To18',
+        'shantytowns.minors_in_school': 'minorsInSchool',
+        'shantytowns.access_to_sanitary': 'accessToSanitary',
+        'shantytowns.sanitary_comments': 'sanitaryComments',
+        'shantytowns.access_to_water': 'accessToWater',
+        'shantytowns.water_comments': 'waterComments',
+        'shantytowns.trash_evacuation': 'trashEvacuation',
+        'shantytowns.owner': 'owner',
+        'shantytowns.census_status::text': 'censusStatus',
+        'shantytowns.census_conducted_by': 'censusConductedBy',
+        'shantytowns.census_conducted_at': 'censusConductedAt',
+        'shantytowns.owner_complaint': 'ownerComplaint',
+        'shantytowns.justice_procedure': 'justiceProcedure',
+        'shantytowns.justice_rendered': 'justiceRendered',
+        'shantytowns.justice_rendered_at': 'justiceRenderedAt',
+        'shantytowns.justice_rendered_by': 'justiceRenderedBy',
+        'shantytowns.justice_challenged': 'justiceChallenged',
+        'shantytowns.police_status::text': 'policeStatus',
+        'shantytowns.police_requested_at': 'policeRequestedAt',
+        'shantytowns.police_granted_at': 'policeGrantedAt',
+        'shantytowns.bailiff': 'bailiff',
+        'shantytowns.closed_with_solutions::text': 'closedWithSolutions',
+        'shantytowns.created_at': 'createdAt',
+        'shantytowns.updated_at': 'updatedAt',
+        // new fields
+        'shantytowns.water_potable': 'waterPotable',
+        'shantytowns.water_continuous_access': 'waterContinuousAccess',
+        'shantytowns.water_public_point': 'waterPublicPoint',
+        'shantytowns.water_distance::text': 'waterDistance',
+        'shantytowns.water_roads_to_cross': 'waterRoadsToCross',
+        'shantytowns.water_everyone_has_access': 'waterEveryoneHasAccess',
+        'shantytowns.water_stagnant_water': 'waterStagnantWater',
+        'shantytowns.water_hand_wash_access': 'waterHandWashAccess',
+        'shantytowns.water_hand_wash_access_number': 'waterHandWashAccessNumber',
+        'shantytowns.sanitary_number': 'sanitaryNumber',
+        'shantytowns.sanitary_insalubrious': 'sanitaryInsalubrious',
+        'shantytowns.sanitary_on_site': 'sanitaryOnSite',
+        'shantytowns.trash_cans_on_site': 'trashCansOnSite',
+        'shantytowns.trash_accumulation': 'trashAccumulation',
+        'shantytowns.trash_evacuation_regular': 'trashEvacuationRegular',
+        'shantytowns.vermin': 'vermin',
+        'shantytowns.vermin_comments': 'verminComments',
+        'shantytowns.fire_prevention_measures': 'firePreventionMeasures',
+        'shantytowns.fire_prevention_diagnostic': 'firePreventionDiagnostic',
+        'shantytowns.fire_prevention_site_accessible': 'firePreventionSiteAccessible',
+        'shantytowns.fire_prevention_devices': 'firePreventionDevices',
+        'shantytowns.fire_prevention_comments': 'firePreventionComments',
+        'shantytowns.resorption_target': 'resorptionTarget',
+        'creators.user_id': 'createdById',
+        'creators.first_name': 'createdByFirstName',
+        'creators.last_name': 'createdByLastName',
+        'creators.position': 'createdByPosition',
+        'creators_organizations.organization_id': 'createdByOrganization',
+        'updators.user_id': 'updatedById',
+        'updators.first_name': 'updatedByFirstName',
+        'updators.last_name': 'updatedByLastName',
+        'updators.position': 'updatedByPosition',
+        'updators_organizations.organization_id': 'updatedByOrganization',
+        'cities.code': 'cityCode',
+        'cities.name': 'cityName',
+        'cities.fk_main': 'cityMain',
+        'cities.latitude': 'cityLatitude',
+        'cities.longitude': 'cityLongitude',
+        'epci.code': 'epciCode',
+        'epci.name': 'epciName',
+        'departements.code': 'departementCode',
+        'departements.name': 'departementName',
+        'departements.latitude': 'departementLatitude',
+        'departements.longitude': 'departementLongitude',
+        // Departement chieftown
+        'departementChiefTown.code': 'departementChiefTownCode',
+        'departementChiefTown.name': 'departementChiefTownName',
+        'departementChiefTown.latitude': 'departementChiefTownLatitude',
+        'departementChiefTown.longitude': 'departementChiefTownLongitude',
+        'regions.code': 'regionCode',
+        'regions.name': 'regionName',
+        'regions.latitude': 'regionLatitude',
+        'regions.longitude': 'regionLongitude',
+        'regionChiefTown.code': 'regionChiefTownCode',
+        'regionChiefTown.name': 'regionChiefTownName',
+        'regionChiefTown.latitude': 'regionChiefTownLatitude',
+        'regionChiefTown.longitude': 'regionChiefTownLongitude',
+        'electricity_types.electricity_type_id': 'electricityTypeId',
+        'electricity_types.label': 'electricityTypeLabel',
+        'shantytowns.electricity_comments': 'electricityComments',
+        'field_types.field_type_id': 'fieldTypeId',
+        'field_types.label': 'fieldTypeLabel',
+        'owner_types.owner_type_id': 'ownerTypeId',
+        'owner_types.label': 'ownerTypeLabel',
+    },
+    joins: [
+        { table: 'owner_types', on: 'shantytowns.fk_owner_type = owner_types.owner_type_id' },
+        { table: 'field_types', on: 'shantytowns.fk_field_type = field_types.field_type_id' },
+        { table: 'electricity_types', on: 'shantytowns.fk_electricity_type = electricity_types.electricity_type_id' },
+        { table: 'cities', on: 'shantytowns.fk_city = cities.code' },
+        { table: 'epci', on: 'cities.fk_epci = epci.code' },
+        { table: 'departements', on: 'cities.fk_departement = departements.code' },
+        { table: 'cities AS departementChiefTown', on: 'departementChiefTown.code = departements.fk_city' },
+        { table: 'regions', on: 'departements.fk_region = regions.code' },
+        { table: 'cities AS regionChiefTown', on: 'regionChiefTown.code = regions.fk_city' },
+        { table: 'users AS creators', on: 'shantytowns.created_by = creators.user_id' },
+        { table: 'organizations AS creators_organizations', on: 'creators.fk_organization = creators_organizations.organization_id' },
+        { table: 'users AS updators', on: 'shantytowns.updated_by = updators.user_id' },
+        { table: 'organizations AS updators_organizations ', on: 'updators.fk_organization = updators_organizations.organization_id' },
+    ],
+};
