+++ packages/api/server/models/shantytownModel/_common/getAddressSimpleOf.js
@@ -1,0 +1,1 @@
+module.exports = shantytown => shantytown.addressSimple || 'Pas d\'adresse précise';
