+++ packages/api/server/models/shantytownModel/_common/serializeComment.js
@@ -1,0 +1,35 @@
+module.exports = comment => Object.assign(
+    {
+        id: comment.commentId,
+        description: comment.commentDescription,
+        createdAt: comment.commentCreatedAt !== null ? (comment.commentCreatedAt.getTime() / 1000) : null,
+        private: comment.commentPrivate,
+        createdBy: {
+            id: comment.commentCreatedBy,
+            first_name: comment.userFirstName,
+            last_name: comment.userLastName,
+            position: comment.userPosition,
+            organization: comment.organizationAbbreviation || comment.organizationName,
+            organization_id: comment.organizationId,
+        },
+        shantytown: comment.shantytownId,
+    },
+    comment.covidCommentDate
+        ? {
+            covid: {
+                date: comment.covidCommentDate.getTime() / 1000,
+                equipe_maraude: comment.covidEquipeMaraude,
+                equipe_sanitaire: comment.covidEquipeSanitaire,
+                equipe_accompagnement: comment.covidEquipeAccompagnement,
+                distribution_alimentaire: comment.covidDistributionAlimentaire,
+                action_mediation_sante: comment.covidActionMediationSante,
+                sensibilisation_vaccination: comment.covidSensibilisationVaccination,
+                equipe_mobile_depistage: comment.covidEquipeMobileDepistage,
+                equipe_mobile_vaccination: comment.covidEquipeMobileVaccination,
+                personnes_orientees: comment.covidPersonnesOrientees,
+                personnes_avec_symptomes: comment.covidPersonnesAvecSymptomes,
+                besoin_action: comment.covidBesoinAction,
+            },
+        }
+        : {},
+);
