+++ packages/api/server/models/shantytownModel/findAll.js
@@ -1,0 +1,3 @@
+const query = require('./_common/query');
+
+module.exports = (user, filters = [], feature = 'list', order = undefined) => query(filters, order, user, feature);
